Add test for add new task and make add new task work
