--- src/Modules/addNew.js
@@ -1,21 +1,0 @@
-// Add new Task
-
-const addNew = (List) => {
-  const desc = document.getElementById('addtodo');
-  if (desc.value) {
-    const completed = false;
-    const description = desc.value;
-    const index = List.length + 1;
-    List.push({ completed, description, index });
-    desc.value = null;
-  }
-  List = List.map((todo, index) => ({
-    completed: todo.completed,
-    description: todo.description,
-    index: index + 1,
-  }));
-
-  return List;
-};
-
-exports.addNew = addNew;
