+++ utils.test.js
@@ -1,4 +1,4 @@
-const { addNew } = require('./src/Modules/addNew.js');
+const { addNew } = require('./src/Modules/utils.js');
 
 beforeEach(() => {
   global.document = {
@@ -14,6 +14,8 @@ afterEach(() => {
 describe('Add new task', () => {
   test('Should return {completed: false, description: "new task", index: 1}', () => {
     document.getElementById.mockReturnValue({ value: 'new task' });
-    expect(addNew([])).toEqual([{completed: false, description: 'new task', index: 1}]);
+    let list = addNew([]);
+    console.log(list);
+    expect(list).toEqual([{completed: false, description: 'new task', index: 1}]);
   });
 });
